translation-pipeline/requirements.txt
flask==3.0.0
gunicorn==21.2.0
anthropic==0.18.1
google-cloud-storage==2.10.0
google-cloud-tasks==2.14.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0
markdown==3.5.1
google-auth==2.23.0
